Everything passes except isPathIgnored itself
