+++ lib/createStylelint.js
@@ -8,6 +8,12 @@ const getConfigForFile = require("./getConfigForFile");
 const getPostcssResult = require("./getPostcssResult");
 const isPathIgnored = require("./isPathIgnored");
 const lintSource = require("./lintSource");
+const fs = require("fs");
+const path = require("path");
+const ignore = require("ignore");
+
+const DEFAULT_IGNORE_FILENAME = ".stylelintignore";
+const FILE_NOT_FOUND_ERROR_CODE = "ENOENT";
 
 // The stylelint "internal API" is passed among functions
 // so that methods on a stylelint instance can invoke
@@ -17,6 +23,24 @@ module.exports = function(
 ) /*: stylelint$internalApi*/ {
   options = options || {};
   const stylelint /*: Object*/ = { _options: options };
+
+  // The ignorer will be used to filter file paths after the glob is checked,
+  // before any files are actually read
+  const ignoreFilePath = options.ignorePath || DEFAULT_IGNORE_FILENAME;
+  const absoluteIgnoreFilePath = path.isAbsolute(ignoreFilePath)
+    ? ignoreFilePath
+    : path.resolve(process.cwd(), ignoreFilePath);
+  let ignoreText = "";
+  try {
+    ignoreText = fs.readFileSync(absoluteIgnoreFilePath, "utf8");
+  } catch (readError) {
+    if (readError.code !== FILE_NOT_FOUND_ERROR_CODE) throw readError;
+  }
+  const ignorePattern = options.ignorePattern || [];
+  stylelint.ignorer = ignore()
+    .add(ignoreText)
+    .add(options.ignoreFiles)
+    .add(ignorePattern);
 
   // Two separate explorers so they can each have their own transform
   // function whose results are cached by cosmiconfig
