+++ lib/standalone.js
@@ -7,14 +7,11 @@ const formatters /*: Object*/ = require("./formatters");
 const fs = require("fs");
 const globby /*: Function*/ = require("globby");
 const hash = require("./utils/hash");
-const ignore = require("ignore");
 const needlessDisables /*: Function*/ = require("./needlessDisables");
 const path = require("path");
 const pify = require("pify");
 const pkg = require("../package.json");
 
-const DEFAULT_IGNORE_FILENAME = ".stylelintignore";
-const FILE_NOT_FOUND_ERROR_CODE = "ENOENT";
 const ALWAYS_IGNORED_GLOBS = ["**/node_modules/**", "**/bower_components/**"];
 
 /*::type CssSyntaxErrorT = {
@@ -54,23 +51,6 @@ module.exports = function(
   let fileCache;
   const startTime = Date.now();
 
-  // The ignorer will be used to filter file paths after the glob is checked,
-  // before any files are actually read
-  const ignoreFilePath = options.ignorePath || DEFAULT_IGNORE_FILENAME;
-  const absoluteIgnoreFilePath = path.isAbsolute(ignoreFilePath)
-    ? ignoreFilePath
-    : path.resolve(process.cwd(), ignoreFilePath);
-  let ignoreText = "";
-  try {
-    ignoreText = fs.readFileSync(absoluteIgnoreFilePath, "utf8");
-  } catch (readError) {
-    if (readError.code !== FILE_NOT_FOUND_ERROR_CODE) throw readError;
-  }
-  const ignorePattern = options.ignorePattern || [];
-  const ignorer = ignore()
-    .add(ignoreText)
-    .add(ignorePattern);
-
   const isValidCode = typeof code === "string";
   if ((!files && !isValidCode) || (files && (code || isValidCode))) {
     throw new Error(
@@ -103,7 +83,10 @@ module.exports = function(
     reportNeedlessDisables,
     syntax,
     customSyntax,
-    fix
+    fix,
+    ignorePath: options.ignorePath,
+    ignorePattern: options.ignorePattern,
+    ignoreFiles: options.ignoreFiles
   });
 
   if (!files) {
@@ -149,8 +132,9 @@ module.exports = function(
         // if file is ignored, return nothing
         if (
           absoluteCodeFilename &&
-          !ignorer.filter(path.relative(process.cwd(), absoluteCodeFilename))
-            .length
+          !stylelint.ignorer.filter(
+            path.relative(process.cwd(), absoluteCodeFilename)
+          ).length
         ) {
           returnValue.output = "";
         } else if (
@@ -189,7 +173,7 @@ module.exports = function(
   return globby(fileList)
     .then(filePaths => {
       // The ignorer filter needs to check paths relative to cwd
-      filePaths = ignorer.filter(
+      filePaths = stylelint.ignorer.filter(
         filePaths.map(p => path.relative(process.cwd(), p))
       );
